Fix the last(?) compression inaccuracy
Finally. For years I've been wondering how to solve that last
inaccuracy. I thought it was impossible: I thought the bug was that
data was being read past the end of the file in memory, which would
have been impossible to recreate since different computers have
different data in their memory.

But while studing Okumura's 1989 LZSS compressor, I realised that it
had the exact same problem Kosinski has: it reads past the end of its
file. Except it doesn't read from invalid memory: it just reads
leftover garbage data in its ring buffer.

By making this compressor use a ring buffer, and allowing it to read
past the end of the file, all of the mismatching files in Sonic 2 are
corrected! All that's left is to flesh-out my test suite with files
from Sonic 3 & Knuckles and confirm that this inaccuracy is fixed for
good, and I can finally put this issue to rest.