Fix a bug in the decompressor
When a byte is written to a memory steam, it's possible for the
internal buffer to be realloc'd, which causes any caches pointers to
become 'stale'. This avoids the use of such a stale pointer, which
was causing garbage data to be written to the decompressed file.

This fixes the remaining non-matching files from Sonic 3 & Knuckles.