Assorted improvements.
It turns out that the Moduled Kosinski comparer was always reporting
that the data was the same, even when it wasn't. Fixing this exposed
that three files in Sonic 3 & Knuckles did not match. This turned out
to be because those three files have invalid headers. These headers
have been corrected so that they can be properly compared.