Make ASM compressor end with newline.
Since the only known copy of an ASM-formatted Kosinski file is
missing its end, we don't know if the original compressor that Sega
used ended its output with a newline or not.

If you're out of the loop, the file I'm referring to can be found
in Sonic the Hedgehog 2's 'Simon Wai' and 'August 21st' prototypes.

https://tcrf.net/Proto:Sonic_the_Hedgehog_2_(Genesis)/Simon_Wai_Prototype#Unassembled_Data
https://tcrf.net/Proto:Sonic_the_Hedgehog_2_(Genesis)/August_21st,_1992#Unassembled_Data